Configure chip as STM32F103xB and set up UART3.
The STM32CubeF1 repo does not contain header files, startup code, or linker scripts for the STM32F103x8. But that's the chip I've got on a Blue Pill module! Using the STM32F103x6 versions instead will work, but will only give us 32k of Flash, 10k RAM, and two UARTs -- the chip actually has 64k Flash, 20k RAM,  and three UARTs. In this change, I've chosen the STM32F103xB variant, which has 128k Flash but the rest of it is the same as the '8' variant (never mind the visual confusion of '8' and 'B'). So we can now configure the third UART on GPIO pins PB10 and PB11 to send 'X' and 'Y' from the main loop. BUT, be careful not to exceed 64k of Flash usage, because the linker will no longer check it for us. We're already close to 32k due to the 'stdio' library ('printf()' in particular).