Set up 3 PWM channels on Timer3.
We want to control RGB LEDs, so we need three digital GPIOs and three PWM pins. GPIOs PB12, PB13, and PB14 are digital signals to drive the first RGB LED. Pins PA6, PA7, and PB0 are set to their Alternate Functions as T3C1, T3C2, and T3C3 for PWM. Timer3 is set up to generate 8-bit PWM for the second RGB LED. TODO: check the frequency of the PWM and use the prescaler if necessary.