WIP: blink LEDs on STM32L1Discovery board.
Note that my Discovery board is a very old one with the STM32L152RBT6 on it (16k RAM, 128k Flash). So I've had to include a linker script locally instead of picking up the one in STM32CubeL1, which is for the newer version of the chip with more RAM and Flash. Not fully functional yet. 500Hz square wave does not seem to work. CPU clock only at 16MHz so far. Blue and green LEDs do blink alternately, though.